requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
